PersistentEffect class update
PersistentEffect updated in accordance with the simplified effect trigger logic.
Also, during execution the PersistentEffect calls for evaluation only on those of its stored effects which managed to choose a target when its selectTarget method was called to prevent another round of target selection which could result in errors.

EFFECTS EMB-8 [IN PROGRESS]
PersistentEffect EMB-13 [IN PROGRESS]