Parametrize target selectors
All implementations of the TargetSelector base class are now parametrized or bound to a certain type parameter.
This has been implemented to allow for parametrization of certain effect types which need to be aware of their target type in order to execute.

TARGETING EMB-10 [IN PROGRESS]
EFFECTS EMB-8 [IN PROGRESS]