Refactor TargetingManager logic and role within effect evaluation
Simplify the TargetingManager's role in the process of selecting targets to improve the clarity of the process and adjust it to the parametrization of Effects and TargetSelectors.
Add appropriate waiting methods to support the VariableTargetSelectors which require player input to choose targets.

TARGETING EMB-10 [IN PROGRESS]
TargetingManager EMB-26 [IN PROGRESS]
VariableTargetSelector EMB-20 [UPDATES NEEDED] -> [IN PROGRESS]
TargetSelector EMB-18 [UPDATES NEEDED] -> [AWAITING TESTING]
ConstantTargetSelector EMB-19 [UPDATES NEEDED] -> [AWAITING TESTING]

EFFECTS EMB-8 [IN PROGRESS]
EffectPlayer EMB-14 [IN PROGRESS]